requirements.txt
streamlit
pypdf
edge-tts
pdf2docx
python-docx
Pillow
streamlit-antd-components
docx2pdf
pywin32
PyMuPDF
moviepy